refactor src folder structure
