+++ src/features/simclr.py
@@ -12,6 +12,7 @@ from tensorflow.keras.applications import ResNet50, ResNet101V2, Xception, Incep
 from tensorflow.keras.preprocessing import image
 from tensorflow.keras.layers import *
 from tensorflow.keras.applications.resnet50 import preprocess_input, decode_predictions
+#from augmentation.gaussian_filter import GaussianBlur
 from utils import *
 import helpers
 import losses
@@ -24,6 +25,25 @@ print(f'Using TensorFlow Version: {tf.__version__}')
 BASE_PATH = './BigEarthData'
 OUTPUT_PATH = os.path.join(BASE_PATH, 'models')
 TFR_PATH = os.path.join(BASE_PATH, 'tfrecords')
+
+def get_negative_mask(batch_size):
+    # return a mask that removes the similarity score of equal/similar images.
+    # this function ensures that only distinct pair of images get their similarity scores
+    # passed as negative examples
+    negative_mask = np.ones((batch_size, 2 * batch_size), dtype=bool)
+    for i in range(batch_size):
+        negative_mask[i, i] = 0
+        negative_mask[i, i + batch_size] = 0
+    return tf.constant(negative_mask)
+
+
+# def gaussian_filter(v1, v2):
+#     k_size = int(v1.shape[1] * 0.1)  # kernel size is set to be 10% of the image height/width
+#     gaussian_ope = GaussianBlur(kernel_size=k_size, min=0.1, max=2.0)
+#     [v1, ] = tf.py_function(gaussian_ope, [v1], [tf.float32])
+#     [v2, ] = tf.py_function(gaussian_ope, [v2], [tf.float32])
+#     return v1, v2
+
 
 # Function for getting a dataset generator for our training data. 
 # The flags affect which tfrecords files to use and how to normalize each.
